==> r_scripts/02_recipes.R <==
# Final Project 2 ----
# Setup pre-processing/recipes

# load packages ----
library(tidyverse)
library(tidymodels)
library(here)

# handle common conflicts
tidymodels_prefer()

set.seed(301)
load(here("data_splits/allies_split.rda"))

# Github link ----
# https://github.com/stat301-2-2024-winter/final-project-2-celenakim



# RECIPE 1: BASIC KITCHEN SINK ----------------------------------------------------------------------
recipe1_kitchen_sink <- recipe(comment_length ~ .,
                          data = allies_train) |> 
  step_rm(comment_id, parent_comment_id, username, comment, word_count, words_per_sentence) |> 
  step_BoxCox(all_numeric_predictors()) |> 
  step_dummy(all_nominal_predictors()) |> 
  step_zv(all_predictors()) |> 
  step_normalize(all_predictors()) 

prep_recipe1_kitchen_sink <- prep(recipe1_kitchen_sink) |> 
  bake(new_data = NULL)

view(prep_recipe1_kitchen_sink)
save(recipe1_kitchen_sink, file = here("recipes/recipe1_kitchen_sink.rda"))


# RECIPE 2: BASIC KITCHEN SINK WITH ONE HOT --------------------------------------------------------------------
recipe2_kitchen_sink_trees <- recipe(comment_length ~ .,
                              data = allies_train) |> 
  step_rm(comment_id, parent_comment_id, username, comment, word_count, words_per_sentence) |> 
  step_dummy(all_nominal_predictors(),
             one_hot = TRUE) |> 
  step_zv(all_predictors()) |> 
  step_normalize(all_predictors()) 

prep_recipe2_kitchen_sink_trees <- prep(recipe2_kitchen_sink_trees) |> 
  bake(new_data = NULL)

view(prep_recipe2_kitchen_sink_trees)
save(recipe2_kitchen_sink_trees, file = here("recipes/recipe2_kitchen_sink_trees.rda"))

# RECIPE 3: RECIPE WITH BOX COX TRANSFORMATIONS AND INTERACTION TERMS ----------------------------------------------------------------------
recipe3_transformed_interactions <- recipe(comment_length ~ focus_present + verb + affiliation + drives + neg_emo + sad + cog_proc + certain + achieve + social,
                              data = allies_train) |> 
  step_BoxCox(all_numeric_predictors()) |> 
  step_dummy(all_nominal_predictors()) |> 
  step_zv(all_predictors()) |> 
  step_normalize(all_predictors()) |> 
  step_interact(terms = ~focus_present:verb) |> 
  step_interact(terms = ~affiliation:drives) |> 
  step_interact(terms = ~neg_emo:sad) |> 
  step_interact(terms = ~cog_proc:certain) |> 
  step_interact(terms = ~achieve:social)

prep_recipe3_transformed_interactions <- prep(recipe3_transformed_interactions) |> 
  bake(new_data = NULL)

view(prep_recipe3_transformed_interactions)
save(recipe3_transformed_interactions, file = here("recipes/recipe3_transformed_interactions.rda"))

# RECIPE 4: RECIPE WITH BOX COX TRANSFORMATIONS AND ONE HOT -------------------------------------------------------
recipe4_transformed_trees <- recipe(comment_length ~ focus_present + verb + affiliation + drives + neg_emo + sad + cog_proc + certain + achieve + social,
                              data = allies_train) |> 
  step_BoxCox(all_numeric_predictors()) |> 
  step_dummy(all_nominal_predictors(),
             one_hot = TRUE) |> 
  step_zv(all_predictors()) |> 
  step_normalize(all_predictors()) 

prep_recipe4_transformed_trees <- prep(recipe4_transformed_trees) |> 
  bake(new_data = NULL)

view(prep_recipe4_transformed_trees)
save(recipe4_transformed_trees, file = here("recipes/recipe4_transformed_trees.rda"))
